requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
Pillow==10.3.0
piexif==1.1.3
smbprotocol==1.14.0
httpx==0.27.0
aiofiles==23.2.1
hachoir==3.3.0
pillow-heif==0.18.0
